Try to make server more stable by running only inside the simulator
Now the CLI is only responsible for installing the test runner and launching it inside the simulator. It appears to be way lighter and more stable than trying to run xcodebuild to install and test. The disadvantage is that I lose access to the logs. But if it really is more stable, so it is a small price to pay.